+++ docs/reference/changelog.rst
@@ -42,6 +42,7 @@ Models
   :py:attr:`~craft_application.AppMetadata.mandatory_adoptable_fields`
   fields of :py:attr:`~craft_application.AppMetadata` must consider
   how these fields should behave in their apps.
+- ``any`` and ``*`` are now reserved names for platforms.
 
 Utilities
 ~~~~~~~~~