+++ docs/reference/changelog.rst
@@ -44,6 +44,13 @@ Models
   how these fields should behave in their apps.
 - ``any`` and ``*`` are now reserved names for platforms.
 
+Fetch Service
+~~~~~~~~~~~~~
+
+- When using the fetch service, the environment variable
+  ``CRAFT_PROVIDERS_EXPERIMENTAL_SUPPRESS_UPGRADE_UNSUPPORTED`` gets set, suppressing
+  running dist-upgrades on instance creation.
+
 Utilities
 ~~~~~~~~~