+++ docs/reference/changelog.rst
@@ -43,6 +43,7 @@ Models
   fields of :py:attr:`~craft_application.AppMetadata` must consider
   how these fields should behave in their apps.
 - ``any`` and ``*`` are now reserved names for platforms.
+- Platform names may not contain '/' characters.
 
 Utilities
 ~~~~~~~~~