+++ docs/reference/changelog.rst
@@ -15,7 +15,7 @@ Changelog
 
     For a complete list of commits, check out the `1.2.3`_ release on GitHub.
 
-6.0.0 (unreleased)
+6.0.0 (2025-11-17)
 ------------------
 
 Models
@@ -62,6 +62,8 @@ Utilities
   location and message.
 - ``format_pydantic_errors`` now takes ``ErrorDetails`` instead of ``ErrorDict``
   objects. In most cases, apps will already be passing ``ErrorDetails`` objects.
+
+For a complete list of commits, check out the `6.0.0`_ release on GitHub.
 
 5.11.0 (2025-10-01)
 -------------------
@@ -1137,3 +1139,4 @@ For a complete list of commits, check out the `2.7.0`_ release on GitHub.
 .. _5.10.2: https://github.com/canonical/craft-application/releases/tag/5.10.2
 .. _5.10.3: https://github.com/canonical/craft-application/releases/tag/5.10.3
 .. _5.11.0: https://github.com/canonical/craft-application/releases/tag/5.11.0
+.. _6.0.0: https://github.com/canonical/craft-application/releases/tag/6.0.0